operator: move certain K8s Node operations to cilium-operator
To decrease the amount of permissions Cilium's requires to operate in a
cluster, the node taint removal and the setup of the node condition
NetworkUnavailable can be set through cilium-operator.

Cilium-operator will remove, if set, the Cilium's specific node taints from
the Kubernetes nodes as well as setting up the NetworkUnavailable node
condition to 'false' once it detects there is a "Ready" Cilium pod in
that node.

Signed-off-by: André Martins <andre@cilium.io>